ProductApi's

+++ ecommerceSolution.ProductsService/BusinessLogicLayer/DTO/ProductAddRequest.cs
@@ -1,7 +1,7 @@
 ﻿
 namespace BusinessLogicLayer.DTO
 {
-    public record ProductAddRequest(string ProductName, CategoryOptions Category, double? UnitPrice, int? QuantityInStock)
+    public record ProductAddRequest(string ProductName, CategoryOptions Category, decimal? UnitPrice, int? QuantityInStock)
     {
         public ProductAddRequest() : this(default, default, default, default)
         {
