+++ ecommerceSolution.ProductsService/BusinessLogicLayer/DTO/ProductResponse.cs
@@ -6,7 +6,7 @@ using System.Threading.Tasks;
 
 namespace BusinessLogicLayer.DTO
 {
-    public record ProductResponse(Guid ProductID, string ProductName, CategoryOptions Category, double? UnitPrice, int? QuantityInStock)
+    public record ProductResponse(Guid ProductID, string ProductName, CategoryOptions Category, decimal? UnitPrice, int? QuantityInStock)
     {
         public ProductResponse() : this(default, default, default, default, default)
         {
