+++ ecommerceSolution.ProductsService/BusinessLogicLayer/DTO/ProductUpdateRequest.cs
@@ -6,7 +6,7 @@ using System.Threading.Tasks;
 
 namespace BusinessLogicLayer.DTO
 {
-    public record ProductUpdateRequest(Guid ProductID, string ProductName, CategoryOptions Category, double? UnitPrice, int? QuantityInStock)
+    public record ProductUpdateRequest(Guid ProductID, string ProductName, CategoryOptions Category, decimal? UnitPrice, int? QuantityInStock)
     {
         public ProductUpdateRequest() : this(default, default, default, default, default)
         {
